changed "id" property of PetClinic's persistent entities from type int to Integer
git-svn-id: https://src.springframework.org/svn/spring-maintenance/trunk@7052 fd5a2b45-1f63-4059-99e9-3c7cb7fd75c8

+++ samples/petclinic/src/org/springframework/samples/petclinic/util/EntityUtils.java
@@ -28,9 +28,10 @@ public abstract class EntityUtils {
 	 */
 	public static Entity getById(Collection entities, Class entityClass, int entityId)
 	    throws ObjectRetrievalFailureException {
+
 		for (Iterator it = entities.iterator(); it.hasNext();) {
 			Entity entity = (Entity) it.next();
-			if (entity.getId() == entityId && entityClass.isInstance(entity)) {
+			if (entity.getId().intValue() == entityId && entityClass.isInstance(entity)) {
 				return entity;
 			}
 		}
